renamed main file to plugin
